Move KafkaConnectorOffsetsAnnotation to io.strimzi.operator.cluster.model
