+++ cluster-operator/src/main/java/io/strimzi/operator/cluster/model/KafkaConnectorOffsetsAnnotation.java
@@ -2,7 +2,7 @@
  * Copyright Strimzi authors.
  * License: Apache License 2.0 (see the file LICENSE or http://apache.org/licenses/LICENSE-2.0.html).
  */
-package io.strimzi.operator.cluster.operator.resource;
+package io.strimzi.operator.cluster.model;
 
 /**
  * Values for strimzi.io/connector-offsets annotation on KafkaConnector or KafkaMirrorMaker2 resource.
